pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "PowerBIMentor"
version = "0.1.0"
description = "Extract metadata from Power BI .pbit templates and evaluate them using AI models."
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Qobiljon Xayrullayev", email = "qobiljonkhayrullayev@gmail.com" }
]
dependencies = [
  "google-genai>=1.0.0",
  "python-dotenv>=1.0.0"
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
exclude = ["tests"]
